One step closer to the original system

+++ src/port/resource/loaders/AudioLoader.h
@@ -8,8 +8,8 @@ extern "C" {
 #endif
 
 char* Audio_LoadBlob(const char* resource, u32 offset);
-AdpcmBook* Audio_LoadBook(uint32_t addr);
-AdpcmLoop* Audio_LoadLoop(uint32_t addr);
+AdpcmBook* Audio_LoadBook(uintptr_t addr);
+AdpcmLoop* Audio_LoadLoop(uintptr_t addr);
 Instrument* Audio_LoadInstrument(uint32_t addr, AudioTableEntry entry, uint32_t sampleBankID);
 Drum* Audio_LoadDrum(uint32_t addr, AudioTableEntry entry, uint32_t sampleBankID);
 Sample* Audio_LoadSample(uint32_t addr, AudioTableEntry entry, uint32_t sampleBankID);
